Split protocol tests into common and dmm
